Turn off all caching of JAR entries as it inhibited reloading FASLs.
The java.net.JarURLConnection implementation *never* invalidates its
cache even for files on the local filesystem, making it highly
unsuitable to represent FASL sources.  One can create a custom
protocol handler, but in the manner oh-so-typical of Java, we cannot
use the default implementation so a sizable amount of coding lies
ahead.  For the time being, we just disable caching.