URL pathnames working for OPEN for built-in schemas.
Still need to decide with URI escaping issues, as we currently rely on
the URL Stream handlers to do the right thing.  And we still need to
retrofit jar pathname's use of a string to represent a URL.

Updates for URL and jar pathname design documents.

Implemented URL-PATHNAME and JAR-PATHNAME as subtypes of PATHNAME.

Adjusted ABCL-TEST-LISP to use functions provided in
"pathname-test.lisp" in "jar-file.lisp".  Added one test for url
pathnames.

Constructor in Java added for a Cons by copying references from the
orignal Cons.