Make JAVA:ADD-TO-CLASSPATH a generic function.
With this change we can customize the mechanism for changing the
classpath.  The first planned use is for JSS to use an :after method
to be informed of classpath additions, so we can factor out the ASDF
portion into the ABCL-ASDF extension package.

Add JAVA:GET-CURRENT-CLASSLOADER to access a wrapped instance of the
underlying current JVM classloader being used by ABCL.